backend/requirements.txt
# Core FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.1

# Configuration
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-xdist==3.6.1
pytest-mock==3.14.0
factory-boy==3.3.1
faker==33.1.0
httpx==0.28.1

# Code Quality
ruff==0.8.4
black==24.10.0
mypy==1.14.1
